+++ components/tetwild/wmtk/components/tetwild/EdgeCollapsing.cpp
@@ -416,8 +416,8 @@ bool TetWildMesh::collapse_edge_after(const Tuple& loc)
 
     // global topology check
     if (m_params.preserve_topology) {
-        std::map<std::pair<size_t, size_t>, int> after_edge_link;
-        std::map<size_t, int> after_vertex_link;
+        std::map<std::pair<size_t, size_t>, size_t> after_edge_link;
+        std::map<size_t, size_t> after_vertex_link;
 
 
         // debug code
